editors/vim: Update to 9.0.1857
On 5 August, 2023, Bram Moolenaar, the long-time maintainer and
architect of Vim, passed away. In addition to being a legendary
software engineer and open-source advocate, he was a major
philanthropist. His impact on the OSS world was immense, as was
his dedication to speaking out for those whose voices are too
often missed.. The Vim project remains in good hands with the
Vim community.